+++ Plugin/CustomPosters/src/Patches/StartOfRoundPatch.cs
@@ -6,6 +6,13 @@ namespace CustomPosters.Patches
     internal class StartOfRoundPatch
     {
         [HarmonyPostfix]
+        [HarmonyPatch("Awake")]
+        private static void OnAwakePatch()
+        {
+            PosterManager.OnSceneAwake();
+        }
+
+        [HarmonyPostfix]
         [HarmonyPatch("Start")]
         private static void OnRoundStartPatch(StartOfRound __instance)
         {